+++ CHANGELOG.md
@@ -1,4 +1,23 @@
-﻿## Version 4.07 [19-03-2026]
+﻿## Version 4.08 [07-04-2026]
+
+**All WoW Versions:**  
+- ElvUI 15.11+ compatibility  
+- Added frFR locales (Credits: Oniria)  
+- Updated CVar values (Nameplate overlap H, Nameplate overlap V)  
+- Updated ElvUI Nameplates (Castbar, Cast target, Buff size, CC icons)  
+- Updated game tooltip default font size (11 > 12)  
+
+**Retail:**  
+- Absorbs color displayed on full health again (Overabsorb)  
+- Added Plumber profile to /lucky config  
+- Updated BuffReminders profile  
+- Updated PermoksAccountManager profile  
+- Updated Private Aura x offset for the latest ElvUI changes (Party frames)  
+
+**Mists:**  
+- Updated TOC for 5.5.3  
+
+## Version 4.07 [19-03-2026]
 
 **All WoW Versions:**  
 - ElvUI 15.09+ compatibility